refactor: handling of evident axios instance
